Project file

+++ 0x1A-hash_tables/6-hash_table_delete.c
@@ -5,7 +5,7 @@
 */
 void hash_table_delete(hash_table_t *ht)
 {
-	int j;
+	unsigned long int j;
 	hash_node_t *holder = NULL;
 	hash_node_t *node = NULL;
